requirements.txt
youtube_transcript_api
google-api-python-client
python-dotenv